% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{compute_SOI}
\alias{compute_SOI}
\title{Core helper to compute a suboptimal/optimal index}
\usage{
compute_SOI(MCD, wCV, rstPct = 20, obsPct = 50, cutoff = 3)
}
\arguments{
\item{MCD}{a mean coverage depth is a the number of genes x the number of samples matrix.}

\item{wCV}{a window coefficient of variation is a the number of genes x the number of samples matrix.}

\item{rstPct}{restricted percent (one-side) to restrict genes by log transformed MC. Default is 20.}

\item{obsPct}{span includes the percent of observations in each local regression. Default is 50.}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}
}
\value{
a matrix including a vector of SOI; a coordinate matrix of smoothed data; and a range of MCD.
}
\description{
Core helper to compute a suboptimal/optimal index
}
\examples{
data("TOY_total")

compute_SOI(
  MCD = TOY_total$MCD,
  wCV = TOY_total$wCV
)
}
