% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{.build_pileupExon}
\alias{.build_pileupExon}
\title{Core helper to build exon-only pileup}
\usage{
.build_pileupExon(pileupPath, cases = NULL, study = NULL)
}
\arguments{
\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{cases}{a vector of specific samples among all samples in pileup. If NULL, all samples are selected. Default is NULL.}

\item{study}{a character of study abbreviation in the pileupList. Default is NULL.}
}
\value{
a numeric matrix of exon-only coverage (rows: exon positions, columns: samples).
}
\description{
Core helper to build exon-only pileup
}
\examples{
## Requires a base-level coverage pileup .RData file
try(
    .build_pileupExon(pileupPath="path/to/pileup.RData"), silent=TRUE
)
}
\references{
Choi, H.Y., Jo, H., Zhao, X. et al. SCISSOR: a framework for identifying structural changes in RNA transcripts. Nat Commun 12, 286 (2021).
}
