% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{filter_lowExpGenes}
\alias{filter_lowExpGenes}
\title{Filter low expressed genes}
\usage{
filter_lowExpGenes(genelist, TPM, thr = 5, pct = 40)
}
\arguments{
\item{genelist}{a vector of gene names}

\item{TPM}{a gene expression counts matrix transformed by TPM}

\item{thr}{threshold. Default is 5.}

\item{pct}{percent. Default is 40.}
}
\value{
a vector of filtered gene names
}
\description{
Filter low expressed genes
}
\examples{
data("TOY_mrna_mat")
filter_lowExpGenes(
  genelist = TOY_mrna_mat$genes,
  TPM      = TOY_mrna_mat$TPM
)
}
