% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{get_pileupExon}
\alias{get_pileupExon}
\title{Get a focused pileup of exon location (for single-study)}
\usage{
get_pileupExon(g, pileupPath, cases = NULL)
}
\arguments{
\item{g}{the gene order in genelist}

\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{cases}{a vector of specific samples among all samples in pileup. If NULL, all samples are selected. Default is NULL.}
}
\value{
a focused pileup is a the number of exon locations x the number of samples matrix for the g-th gene.
}
\description{
Get a focused pileup of exon location (for single-study)
}
\examples{
## Requires a base-level coverage pileup .RData file
try(
    get_pileupExon(g=1, pileupPath=c("path/to/pileup1.RData", "path/to/pileup2.RData")), silent=TRUE
)
}
