% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{get_SOI}
\alias{get_SOI}
\title{Get a suboptimal/optimal index for samples}
\usage{
get_SOI(
  MCD,
  wCV = NULL,
  rstPct = 20,
  obsPct = 50,
  cutoff = 3,
  assay.MCD = "MCD",
  assay.wCV = "wCV"
)
}
\arguments{
\item{MCD}{a mean coverage depth is a the number of genes x the number of samples matrix.}

\item{wCV}{a window coefficient of variation is a the number of genes x the number of samples matrix.}

\item{rstPct}{restricted percent (one-side) to restrict genes by log transformed MC. Default is 20.}

\item{obsPct}{span includes the percent of observations in each local regression. Default is 50.}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}

\item{assay.MCD}{character string specifying the assay name containing the MCD matrix in a SummarizedExperiment object.}

\item{assay.wCV}{character string specifying the assay name containing the wCV matrix in a SummarizedExperiment object.}
}
\value{
a matrix including a vector of SOI; a coordinate matrix of smoothed data; and a range of MCD.
}
\description{
Get a suboptimal/optimal index for samples
}
\examples{
data("TOY_total_se")

get_SOI(TOY_total_se)
}
