% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shape.R
\name{plot_GBC}
\alias{plot_GBC}
\title{Plot gene body coverage}
\usage{
plot_GBC(
  pileupPath,
  geneNames,
  rnum = 100,
  method = 1,
  scale = TRUE,
  stat = 2,
  plot = TRUE,
  sampleInfo
)
}
\arguments{
\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{geneNames}{gene names per file. If NULL, Gene i with the same length of pileupPath be set. Default is NULL.}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}

\item{scale}{TRUE/FALSE returns the scaled/unscaled normalized transcript coverage. Default is TRUE.}

\item{stat}{1 and 2 return median and mean normalized coverage curves per sample, respectively. Default is 1.}

\item{plot}{TRUE/FALSE turns on/off the normalized transcript coverage plot. Default is TRUE.}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}
}
\value{
a matrix and a plot, or a matrix for the gene body coverage where plot is TRUE or FALSE, respectively.
}
\description{
Plot gene body coverage
}
\examples{
## Interface-only example
try(
    plot_GBC(
        pileupPath = NA,
        geneNames  = "GENE1",
        sampleInfo = data.frame(
            SampleID       = c("S1", "S2"),
            CODING_BASES   = c(1, 1),
            INTRONIC_BASES = c(1, 1)
        ),
        plot       = FALSE
    ),
    silent=TRUE
)
}
