% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shape.R
\name{plot_SOI}
\alias{plot_SOI}
\title{Plot suboptimal/optimal index outputs}
\usage{
plot_SOI(SOIresult, cutoff = 3, outFile = NULL)
}
\arguments{
\item{SOIresult}{outputs from \code{get_SOI} function}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}

\item{outFile}{a directory with a file name to save outputs. Default is NULL.}
}
\value{
figures for the distribution of SOI by PD; and the relation of wCV and MCD.
}
\description{
Plot suboptimal/optimal index outputs
}
\examples{
data("TOY_total_se")

res <- get_SOI(TOY_total_se)

plot_SOI(SOIresult=res, outFile=tempfile(fileext=".png"))
}
\references{
https://jtr13.github.io/cc21fall2/raincloud-plot-101-density-plot-or-boxplotwhy-not-do-both.html
}
