% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{combine_vecObj}
\alias{combine_vecObj}
\title{Combine vectors as a matrix from objects}
\usage{
combine_vecObj(
  filePath,
  objName = NULL,
  header = NULL,
  skip = NULL,
  txtCol = NULL,
  margin,
  rowNames,
  colNames,
  nCores = 32
)
}
\arguments{
\item{filePath}{file paths including .RData file names}

\item{objName}{a character of object name}

\item{header}{logical; whether the text files have a header line.}

\item{skip}{integer; number of lines to skip before reading data from `.txt` files.}

\item{txtCol}{integer; column index in the text file that contains the
numeric vector to be extracted.}

\item{margin}{1 and 2 return for gene- and sample-level vectors, respectively.}

\item{rowNames}{a vector of gene names}

\item{colNames}{a vector of sample names}

\item{nCores}{the number of cores for parallel computing. Default is 32.}
}
\value{
a gene x sample matrix
}
\description{
Combine vectors as a matrix from objects
}
\examples{
\dontrun{
DR.mat <- combine_vecObj(
  filePath = paste0("path/to/", genelist, "/", Study, "_", genelist, "_degradation.RData"),
  objName  = "allSampleDegRate",
  margin   = 1,
  rowNames = genelist,
  colNames = cases,
  nCores   = 32
)
}
}
