% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{compute_DIIwt}
\alias{compute_DIIwt}
\title{Core helper to compute a degraded/intact index using gene weight}
\usage{
compute_DIIwt(DR, alpha = 2, cutoff = 3, TPM, thr = 5, pct = 40, genelength)
}
\arguments{
\item{DR}{a the number of genes x the number of samples matrix of decay rates}

\item{alpha}{a positive numeric exponent factor to weight the magnitude of decay rates. Default is 2.}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}

\item{TPM}{a numeric matrix of TPM values with the same genes in rows and the same samples in columns as \code{DR}.}

\item{thr}{threshold. Default is 5.}

\item{pct}{percent. Default is 40.}

\item{genelength}{a gene length (bp) vector with names as gene IDs.}
}
\value{
a matrix of with decay rate with filtered genes; a matrix including a vector of DII; a data frame of gene info; and a scale factor.
}
\description{
Core helper to compute a degraded/intact index using gene weight
}
\examples{
data("TOY_mrna")

compute_DIIwt(
  DR         = TOY_mrna$DR,
  TPM        = TOY_mrna$TPM,
  genelength = TOY_mrna$genelength
)
}
