% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{compute_MCD}
\alias{compute_MCD}
\title{Core helper to compute a mean coverage depth}
\usage{
compute_MCD(pileupData, sampleInfo, cases = NULL)
}
\arguments{
\item{pileupData}{exon-only coverage pileup matrix for a single gene.}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}

\item{cases}{optional character vector specifying a subset of samples.
used for handling missing coverage.}
}
\description{
Core helper to compute a mean coverage depth
}
