% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{compute_wCV}
\alias{compute_wCV}
\title{Core helper to compute a window coefficient of variation}
\usage{
compute_wCV(
  pileupData,
  sampleInfo,
  rnum = 100,
  method = 1,
  winSize = 20,
  egPct = 10,
  cases = NULL
)
}
\arguments{
\item{pileupData}{exon-only coverage pileup matrix for a single gene.}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}

\item{winSize}{window size of the rolling window. Default is 20.}

\item{egPct}{edge percent (one-side) to calculate the trimmed mean. Default is 10.}

\item{cases}{optional character vector specifying a subset of samples.
used for handling missing coverage.}
}
\description{
Core helper to compute a window coefficient of variation
}
