% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_RData}
\alias{extract_RData}
\title{Extract an object from .RData}
\usage{
extract_RData(file, object)
}
\arguments{
\item{file}{.RData file}

\item{object}{object name}
}
\value{
the object
}
\description{
Extract an object from .RData
}
\references{
https://stackoverflow.com/questions/65964064/programmatically-extract-an-object-from-collection-of-rdata-files
}
