% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{get_DIIhc}
\alias{get_DIIhc}
\title{Get a degraded/intact index for samples using hierarchical clustering}
\usage{
get_DIIhc(DR, topPct = 5)
}
\arguments{
\item{DR}{a the number of genes x the number of samples matrix of decay rates}

\item{topPct}{top percentages of decay rates defined as degrateGrp=1. Default is 5.}
}
\value{
a matrix of binary converted decay rates; hierarchical clustering outputs of samples; and a vector of DII per sample.
}
\description{
Get a degraded/intact index for samples using hierarchical clustering
}
\examples{
data("TOY_mrna")
get_DIIhc(DR = TOY_mrna$DR)
}
