% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{get_DIIwt}
\alias{get_DIIwt}
\title{Get a degraded/intact index for samples using gene weight}
\usage{
get_DIIwt(DR, alpha = 1.5, cutoff = 3, TPM, thru = 5, pct = 40, genelength.mat)
}
\arguments{
\item{DR}{a the number of genes x the number of samples matrix of decay rates}

\item{alpha}{a positive numeric exponent factor to weight the magnitude of decay rates. Default is 1.5.}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}

\item{TPM}{a numeric matrix of TPM values with the same genes in rows and the same samples in columns as \code{DR}.}

\item{thru}{threshold. Default is 5.}

\item{pct}{percent. Default is 40.}

\item{genelength.mat}{a one-column gene length (bp) matrix with row names as gene IDs (no column names).}
}
\value{
a matrix of with decay rate with filtered genes; a matrix including a vector of DII; a data frame of gene info; and a scale factor.
}
\description{
Get a degraded/intact index for samples using gene weight
}
