% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{get_DR}
\alias{get_DR}
\title{Get a decay rate for genes and samples (for a genelist)}
\usage{
get_DR(genelist, pileupPath, sampleInfo, cases = NULL, nCores = 2)
}
\arguments{
\item{genelist}{a vector of gene names}

\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}

\item{cases}{a vector of specific samples among all samples in pileup. If NULL, all samples are selected. Default is NULL.}

\item{nCores}{the number of cores for parallel computing. Default is 2.}
}
\value{
DR is a the number of genes x the number of samples matrix.
}
\description{
Get a decay rate for genes and samples (for a genelist)
}
\examples{
## NOTE:
## This example demonstrates the function interface only.
## Meaningful results require coverage pileup files generated
## from BAM files (see vignette for a full workflow).
data("TOY_mrna_mat")

## Interface-only example (no meaningful output is produced)
try(
    get_DR(
        genelist   = TOY_mrna_mat$genes,
        pileupPath = rep(NA, length(TOY_mrna_mat$genes)),
        sampleInfo = data.frame(SampleID=TOY_mrna_mat$samples),
        nCores     = 2
    ),
    silent=TRUE
)
}
\references{
Choi, H.Y., Jo, H., Zhao, X. et al. SCISSOR: a framework for identifying structural changes in RNA transcripts. Nat Commun 12, 286 (2021).
}
