% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC.R
\name{get_DR}
\alias{get_DR}
\title{Get a decay rate for genes and samples (for a genelist)}
\usage{
get_DR(genelist, pileupPath, sampleInfo, cases = NULL, nCores = 32)
}
\arguments{
\item{genelist}{a vector of gene names}

\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}

\item{cases}{a vector of specific samples among all samples in pileup. If NULL, all samples are selected. Default is NULL.}

\item{nCores}{the number of cores for parallel computing. Default is 32.}
}
\value{
DR is a the number of genes x the number of samples matrix.
}
\description{
Get a decay rate for genes and samples (for a genelist)
}
\references{
Choi, H.Y., Jo, H., Zhao, X. et al. SCISSOR: a framework for identifying structural changes in RNA transcripts. Nat Commun 12, 286 (2021).
}
