% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shape.R
\name{plot_DIIwt}
\alias{plot_DIIwt}
\title{Plot degraded/intact index outputs}
\usage{
plot_DIIwt(DR, DIIresult, cutoff = 3, outFile = NULL)
}
\arguments{
\item{DR}{a the number of genes x the number of samples matrix of decay rates}

\item{DIIresult}{outputs from \code{get_DII} function}

\item{cutoff}{numeric threshold on projection depth used to classify samples.}

\item{outFile}{a directory with a file name to save outputs. Default is NULL.}
}
\value{
figures for the distribution of DII by PD; and the heatmap of DR.
}
\description{
Plot degraded/intact index outputs
}
\examples{
data("TOY_mrna")

res <- get_DIIwt(
  DR             = TOY_mrna$DR,
  TPM            = TOY_mrna$TPM,
  genelength.mat = TOY_mrna$genelength.mat
)

try(
  plot_DIIwt(
    DR        = TOY_mrna$DR,
    DIIresult = res,
    outFile   = tempfile(fileext=".png")
  ),
  silent = TRUE
)
}
\references{
https://jtr13.github.io/cc21fall2/raincloud-plot-101-density-plot-or-boxplotwhy-not-do-both.html
}
