% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shape.R
\name{plot_GBCos}
\alias{plot_GBCos}
\title{Plot gene body coverage with optimal samples}
\usage{
plot_GBCos(stat = 2, plot = TRUE, sampleInfo, GBCresult, auc.vec)
}
\arguments{
\item{stat}{1 and 2 return median and mean normalized coverage curves per sample, respectively. Default is 1.}

\item{plot}{TRUE/FALSE turns on/off the normalized transcript coverage plot. Default is TRUE.}

\item{sampleInfo}{a sample information table including sample id. The number of rows is equal to the number of samples.}

\item{GBCresult}{results of the gene body coverage with all samples}

\item{auc.vec}{a vector with SOI per sample}
}
\value{
a matrix and a plot, or a matrix for the gene body coverage where plot is TRUE or FALSE, respectively.
}
\description{
Plot gene body coverage with optimal samples
}
\examples{
## Interface-only example
GBCresult <- list(
  GBP = data.frame(
    region = 1:2,
    sample = c("S1", "S2"),
    scale.geom = c(1, 1),
    RatioIntron = c(1, 1)
  )
)

auc.vec <- data.frame(
  Sample = c("S1", "S2"),
  PD = c(0, 0),
  SOI = c("Optimal", "Optimal")
)

try(
  plot_GBCos(
    sampleInfo = data.frame(SampleID=c("S1","S2")),
    GBCresult  = GBCresult,
    auc.vec    = auc.vec,
    plot       = FALSE
  ),
  silent = TRUE
)
}
